requirements.txt
requests
python-dotenv
tabulate
questionary
rich
